pyproject.toml
[tool.poetry]
name = "chada-vit"
version = "0.1.0"
description = "Channel Adaptive Vision Transformer repository"
authors = ["Nicolas Bourriez <nicolas.bourriez@gmail.com>"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pytorch-lightning = "2.0.2"
scipy = "^1.11.4"
absl-py = "1.4.0"
accelerate = "0.20.3"
addict = "2.4.0"
aiohttp = "3.8.4"
aiosignal = "1.3.1"
albumentations = "1.3.1"
aliyun-python-sdk-core = "2.13.36"
aliyun-python-sdk-kms = "2.16.1"
altair = "4.2.2"
antlr4-python3-runtime = "4.9.3"
appdirs = "1.4.4"
argcomplete = "1.12.0"
asttokens = "2.0.5"
async-timeout = "4.0.2"
attrs = "23.1.0"
backcall = "0.2.0"
bleach = "4.1.0"
blinker = "1.6.2"
bokeh = "3.1.1"
bottleneck = "1.3.5"
brotlipy = "0.7.0"
cachetools = "5.3.0"
cchardet = "2.1.7"
certifi = "2023.5.7"
cffi = "1.15.1"
cfgv = "3.3.1"
chardet = "4.0.0"
charset-normalizer = "2.0.4"
click = "8.0.4"
cloudpickle = "2.2.1"
colorama = "0.4.6"
colorcet = "3.0.1"
comm = "0.1.2"
contourpy = "1.0.5"
crcmod = "1.7"
cryptography = "39.0.1"
cycler = "0.11.0"
cytoolz = "0.12.0"
dask = "2023.6.0"
datasets = "2.13.1"
datashader = "0.15.1"
datashape = "0.5.2"
debugpy = "1.5.1"
decorator = "5.1.1"
dill = "0.3.6"
distlib = "0.3.6"
docker-pycreds = "0.4.0"
easydict = "1.10"
einops = "0.6.1"
entrypoints = "0.4"
evaluate = "0.4.0"
executing = "0.8.3"
filelock = "3.12.2"
fonttools = "4.25.0"
frozenlist = "1.3.3"
fsspec = "2023.4.0"
future = "0.18.3"
gitdb = "4.0.10"
gitpython = "3.1.31"
google-auth = "2.20.0"
google-auth-oauthlib = "1.0.0"
grpcio = "1.54.2"
h5py = "3.9.0"
holoviews = "1.16.2"
huggingface-hub = "0.15.1"
hydra-core = "1.3.2"
identify = "2.5.24"
idna = "3.4"
idr-py = "0.4.2"
ijson = "3.2.0.post0"
imagecodecs = "2021.8.26"
imageio = "2.31.0"
importlib-metadata = "6.8.0"
ipykernel = "6.19.2"
ipython = "8.12.0"
ipywidgets = "8.0.4"
jedi = "0.18.1"
jinja2 = "3.1.2"
jmespath = "0.10.0"
joblib = "1.2.0"
jsonschema = "4.17.3"
jupyter-client = "8.1.0"
jupyter-core = "5.3.0"
jupyterlab-widgets = "3.0.5"
kiwisolver = "1.4.4"
lazy-loader = "0.2"
lightning-utilities = "0.9.0"
linkify-it-py = "2.0.0"
llvmlite = "0.40.0"
locket = "1.0.0"
lpips = "0.1.4"
markdown = "3.4.1"
markdown-it-py = "2.2.0"
markupsafe = "2.1.1"
matplotlib = "3.7.1"
matplotlib-inline = "0.1.6"
mmsegmentation = "1.1.1"
model-index = "0.1.11"
multidict = "6.0.4"
multipledispatch = "0.6.0"
multiprocess = "0.70.14"
munkres = "1.1.4"
nest-asyncio = "1.5.6"
networkx = "2.8.4"
nodeenv = "1.8.0"
numba = "0.57.0"
numexpr = "2.8.4"
numpy = "1.24.3"
oauthlib = "3.2.2"
omegaconf = "2.3.0"
omero-py = "5.7.1"
opencv-python = "4.7.0.72"
opencv-python-headless = "4.8.0.74"
opendatalab = "0.0.10"
openmim = "0.3.9"
openxlab = "0.0.17"
ordered-set = "4.1.0"
oss2 = "2.17.0"
packaging = "23.0"
pandas = "1.5.3"
panel = "1.1.0"
param = "1.13.0"
parso = "0.8.3"
partd = "1.2.0"
pathos = "0.3.0"
pathtools = "0.1.2"
pexpect = "4.8.0"
pickleshare = "0.7.5"
pillow = "9.4.0"
pip = "23.1.2"
pipx = "1.2.0"
platformdirs = "3.8.1"
ply = "3.11"
pooch = "1.4.0"
pox = "0.3.2"
ppft = "1.7.6.6"
pre-commit = "3.3.3"
prettytable = "3.8.0"
prompt-toolkit = "3.0.36"
protobuf = "3.20.3"
psutil = "5.9.0"
ptyprocess = "0.7.0"
pure-eval = "0.2.2"
pyarrow = "12.0.0"
pyasn1 = "0.5.0"
pyasn1-modules = "0.3.0"
pycparser = "2.21"
pycryptodome = "3.18.0"
pyct = "0.5.0"
pydeck = "0.8.1b0"
pydeprecate = "0.3.2"
pygments = "2.15.1"
pympler = "1.0.1"
pynndescent = "0.5.10"
pyopenssl = "23.0.0"
pyparsing = "3.0.9"
pyqt5-sip = "12.11.0"
pyrsistent = "0.19.3"
pysocks = "1.7.1"
python-dateutil = "2.8.2"
pytz = "2023.3"
pyviz-comms = "2.3.0"
pywavelets = "1.4.1"
pyyaml = "6.0"
pyzmq = "25.1.0"
qudida = "0.0.4"
requests = "2.28.2"
requests-oauthlib = "1.3.1"
responses = "0.18.0"
rich = "13.4.2"
rsa = "4.9"
safetensors = "0.3.1"
scikit-image = "0.19.2"
scikit-learn = "1.2.2"
seaborn = "0.12.2"
sentry-sdk = "1.25.1"
setproctitle = "1.3.2"
setuptools = "60.2.0"
sip = "6.6.2"
six = "1.16.0"
smmap = "5.0.0"
stack-data = "0.2.0"
streamlit = "1.22.0"
submitit = "1.4.5"
tabulate = "0.9.0"
tenacity = "8.2.2"
tensorboard = "2.13.0"
tensorboard-data-server = "0.7.1"
termcolor = "2.3.0"
threadpoolctl = "3.1.0"
tifffile = "2021.7.2"
timm = {git = "https://github.com/huggingface/pytorch-image-models.git", rev = "fc5d705b83fef6f89c7c090903a3ff6fbd7fd0dd"}
toml = "0.10.2"
tomli = "2.0.1"
toolz = "0.12.0"
torch = "2.0.1"
torchaudio = "2.0.2"
torchmetrics = "0.10.0"
torchvision = "0.15.2"
tornado = "6.2"
tqdm = "4.65.0"
traitlets = "5.7.1"
typing-extensions = "4.6.3"
tzlocal = "5.0.1"
uc-micro-py = "1.0.1"
umap-learn = "0.5.3"
urllib3 = "1.26.16"
userpath = "1.7.0"
validators = "0.20.0"
virtualenv = "20.23.1"
wandb = "0.15.4"
watchdog = "3.0.0"
wcwidth = "0.2.5"
webencodings = "0.5.1"
werkzeug = "2.3.6"
wheel = "0.38.4"
widgetsnbextension = "4.0.5"
xarray = "2023.5.0"
xxhash = "3.2.0"
xyzservices = "2022.9.0"
yapf = "0.40.1"
yarl = "1.9.2"
zeroc-ice = "3.6.5"
zipp = "3.11.0"

[tool.poetry.group.dev.dependencies]
accelerate = "0.20.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
